+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Added option "Both direction" into "Expand to".
+* Update button grab system.